requirements.txt
boto3==1.35.60
python-dotenv==1.0.1
flask
